refactor: consolidate InvalidOperationKindError into @taquito/core and improve across packages
re #1994
