+++ packages/taquito-utils/src/errors.ts
@@ -10,6 +10,7 @@ export {
   InvalidChainIdError,
   InvalidKeyHashError,
   InvalidOperationHashError,
+  InvalidOperationKindError,
 } from '@taquito/core';
 
 /**
@@ -20,17 +21,6 @@ export class InvalidProtocolHashError extends Error {
   public name = 'InvalidProtocolHashError';
   constructor(public protocolHash: string) {
     super(`The protocol hash '${protocolHash}' is invalid`);
-  }
-}
-
-/**
- *  @category Error
- *  @description Error that indicates an invalid operation kind being passed or used
- */
-export class InvalidOperationKindError extends Error {
-  public name = 'InvalidOperationKindError';
-  constructor(public operationKind: string) {
-    super(`The operation kind '${operationKind}' is unsupported`);
   }
 }
